Implement new layout on profile page
