+++ components/index.ts
@@ -1,4 +1,4 @@
-export { Layout } from './Layout';
+export { Layout } from './Layout/Layout';
 export { Button } from './Button';
 export { Avatar } from './Avatar';
 export { Pill } from './Pill';
